Use the wildmatch lib instead of fnmatch(3) for globstar support
The library is added as a git submodule and a Meson subproject, because
there has been no upstream release since 2016, and the upstream CMake
build system does not install things such as a pkg-config file or a
FindWildmatch.cmake file.

If that were to change, Meson should automatically prefer the pkg-config
or CMake package over the bundled subproject.